backend/requirements.txt
Flask
flask-cors
bcrypt
PyJWT
pymongo
